Fix various minor issues.
This removes the extraneous leading length field in the signals object, as
per new format spec.

This also puts the events and alarms objects in the namespace object into
the proper order.

This also adds tests for equality to the alarm object test suite.

This also fixes incorrect accounting of sizes when including size fields.