Modify ruby generator to use an implicit null terminator for strings.
The ruby generator used to use the string's length (without the null
terminator) and padded that to the nearest multiple of 4 bytes, but we
should ensure that the string doesn't contain any null bytes, then
implicitly add one using `pack`.

Also break part of push_str into push_buffer, because, now, not all
buffers are strings, but all strings are buffers.